+++ packages/tailwindcss/src/feature-flags.ts
@@ -1,0 +1,1 @@
+export const enableRemoveUnusedThemeVariables = process.env.FEATURES_ENV !== 'stable'