+++ packages/tailwindcss/src/compat/screens-config.test.ts
@@ -46,7 +46,13 @@ test('CSS `--breakpoint-*` merge with JS config `screens`', async () => {
       'print:items-end',
     ]),
   ).toMatchInlineSnapshot(`
-    ".max-w-screen-sm {
+    ":root, :host {
+      --breakpoint-md: 50rem;
+      --breakpoint-lg: 64rem;
+      --breakpoint-xl: 80rem;
+      --breakpoint-2xl: 96rem;
+    }
+    .max-w-screen-sm {
       max-width: 44rem;
     }
     .sm\\:flex {
@@ -134,7 +140,10 @@ test('JS config `screens` extend CSS `--breakpoint-*`', async () => {
       'print:items-end',
     ]),
   ).toMatchInlineSnapshot(`
-    ".min-xs\\:flex {
+    ":root, :host {
+      --breakpoint-md: 50rem;
+    }
+    .min-xs\\:flex {
       @media (width >= 30rem) {
         display: flex;
       }
@@ -307,7 +316,14 @@ test('JS config `screens` overwrite CSS `--breakpoint-*`', async () => {
       'print:items-end',
     ]),
   ).toMatchInlineSnapshot(`
-    ".mini\\:flex {
+    ":root, :host {
+      --breakpoint-sm: 40rem;
+      --breakpoint-md: 48rem;
+      --breakpoint-lg: 64rem;
+      --breakpoint-xl: 80rem;
+      --breakpoint-2xl: 96rem;
+    }
+    .mini\\:flex {
       @media (width >= 40rem) {
         display: flex;
       }
@@ -568,7 +584,10 @@ describe('complex screen configs', () => {
         'print:items-end',
       ]),
     ).toMatchInlineSnapshot(`
-      ".min-sm\\:flex {
+      ":root, :host {
+        --breakpoint-md: 48rem;
+      }
+      .min-sm\\:flex {
         @media (width >= 40rem) {
           display: flex;
         }