+++ packages/tailwindcss/src/utilities.test.ts
@@ -147,6 +147,7 @@ test('inset', async () => {
   ).toMatchInlineSnapshot(`
     ":root, :host {
       --spacing-4: 1rem;
+      --inset-shadow-sm: inset 0 1px 1px #0000000d;
       --inset-shadowned: 1940px;
     }
 
@@ -3161,6 +3162,10 @@ describe('container', () => {
     ).toMatchInlineSnapshot(`
       ":root, :host {
         --breakpoint-sm: 40rem;
+        --breakpoint-md: 48rem;
+        --breakpoint-lg: 64rem;
+        --breakpoint-xl: 80rem;
+        --breakpoint-2xl: 96rem;
       }
 
       .container {
@@ -3300,6 +3305,10 @@ describe('container', () => {
     ).toMatchInlineSnapshot(`
       ":root, :host {
         --breakpoint-sm: 40rem;
+        --breakpoint-md: 48rem;
+        --breakpoint-lg: 64rem;
+        --breakpoint-xl: 80rem;
+        --breakpoint-2xl: 96rem;
       }
 
       .container {
@@ -4036,7 +4045,11 @@ test('translate-x', async () => {
       ['translate-x-full', '-translate-x-full', 'translate-x-px', '-translate-x-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ".-translate-x-\\[var\\(--value\\)\\] {
+    ":root, :host {
+      --spacing: .25rem;
+    }
+
+    .-translate-x-\\[var\\(--value\\)\\] {
       --tw-translate-x: calc(var(--value) * -1);
       translate: var(--tw-translate-x) var(--tw-translate-y);
     }
@@ -4160,7 +4173,11 @@ test('translate-y', async () => {
       ['translate-y-full', '-translate-y-full', 'translate-y-px', '-translate-y-[var(--value)]'],
     ),
   ).toMatchInlineSnapshot(`
-    ".-translate-y-\\[var\\(--value\\)\\] {
+    ":root, :host {
+      --spacing: .25rem;
+    }
+
+    .-translate-y-\\[var\\(--value\\)\\] {
       --tw-translate-y: calc(var(--value) * -1);
       translate: var(--tw-translate-x) var(--tw-translate-y);
     }
@@ -7717,7 +7734,11 @@ test('divide-x with custom default border width', async () => {
       ['divide-x'],
     ),
   ).toMatchInlineSnapshot(`
-    ":where(.divide-x > :not(:last-child)) {
+    ":root, :host {
+      --default-border-width: 2px;
+    }
+
+    :where(.divide-x > :not(:last-child)) {
       --tw-divide-x-reverse: 0;
       border-inline-style: var(--tw-border-style);
       border-inline-start-width: calc(2px * var(--tw-divide-x-reverse));
@@ -7819,7 +7840,11 @@ test('divide-y with custom default border width', async () => {
       ['divide-y'],
     ),
   ).toMatchInlineSnapshot(`
-    ":where(.divide-y > :not(:last-child)) {
+    ":root, :host {
+      --default-border-width: 2px;
+    }
+
+    :where(.divide-y > :not(:last-child)) {
       --tw-divide-y-reverse: 0;
       border-bottom-style: var(--tw-border-style);
       border-top-style: var(--tw-border-style);
@@ -9905,7 +9930,11 @@ test('border with custom default border width', async () => {
       ['border'],
     ),
   ).toMatchInlineSnapshot(`
-    ".border {
+    ":root, :host {
+      --default-border-width: 2px;
+    }
+
+    .border {
       border-style: var(--tw-border-style);
       border-width: 2px;
     }
@@ -13844,6 +13873,7 @@ test('transition', async () => {
     ":root, :host {
       --default-transition-timing-function: ease;
       --default-transition-duration: .1s;
+      --transition-property: color, background-color, border-color, text-decoration-color, fill, stroke, opacity, box-shadow, transform, filter, backdrop-filter;
       --transition-property-opacity: opacity;
     }
 
@@ -13909,7 +13939,12 @@ test('transition', async () => {
       ['transition', 'transition-all', 'transition-colors'],
     ),
   ).toMatchInlineSnapshot(`
-    ".transition {
+    ":root, :host {
+      --default-transition-timing-function: ease;
+      --default-transition-duration: .1s;
+    }
+
+    .transition {
       transition-property: color, background-color, border-color, outline-color, text-decoration-color, fill, stroke, --tw-gradient-from, --tw-gradient-via, --tw-gradient-to, opacity, box-shadow, transform, translate, scale, rotate, filter, -webkit-backdrop-filter, backdrop-filter;
       transition-timing-function: var(--tw-ease, ease);
       transition-duration: var(--tw-duration, .1s);
@@ -15195,6 +15230,8 @@ test('shadow', async () => {
   ).toMatchInlineSnapshot(`
     ":root, :host {
       --color-red-500: #ef4444;
+      --shadow-sm: 0 1px 3px 0 #0000001a, 0 1px 2px -1px #0000001a;
+      --shadow-xl: 0 20px 25px -5px #0000001a, 0 8px 10px -6px #0000001a;
     }
 
     .shadow-\\[10px_10px\\] {
@@ -15417,6 +15454,8 @@ test('inset-shadow', async () => {
   ).toMatchInlineSnapshot(`
     ":root, :host {
       --color-red-500: #ef4444;
+      --inset-shadow: inset 0 2px 4px #0000000d;
+      --inset-shadow-sm: inset 0 1px 1px #0000000d;
     }
 
     .inset-shadow {
@@ -16407,6 +16446,7 @@ describe('spacing utilities', () => {
     expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
       ":root, :host {
         --spacing-sm: 8px;
+        --container-sm: 256px;
       }
 
       .w-sm {
@@ -16480,7 +16520,13 @@ describe('custom utilities', () => {
     let compiled = build([])
 
     // `foo` is not used yet:
-    expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`"@layer utilities;"`)
+    expect(optimizeCss(compiled).trim()).toMatchInlineSnapshot(`
+      "@layer utilities;
+
+      :root, :host {
+        --example-foo: 123px;
+      }"
+    `)
 
     // `foo` is used, and the CSS variable is emitted:
     compiled = build(['foo'])
@@ -17664,7 +17710,12 @@ describe('custom utilities', () => {
       `
 
       expect(await compileCss(input, ['flex'])).toMatchInlineSnapshot(`
-        ".flex {
+        ":root, :host {
+          --example-foo: red;
+          --color-red-500: red;
+        }
+
+        .flex {
           display: flex;
         }"
       `)
@@ -17757,7 +17808,11 @@ describe('custom utilities', () => {
     `
 
     expect(await compileCss(input, ['tab-github'])).toMatchInlineSnapshot(`
-      ".tab-github {
+      ":root, :host {
+        --tab-size-github: 8;
+      }
+
+      .tab-github {
         tab-size: 8;
       }"
     `)