+++ packages/tailwindcss/src/variants.test.ts
@@ -747,7 +747,15 @@ test('default breakpoints', async () => {
       ['sm:flex', 'md:flex', 'lg:flex', 'xl:flex', '2xl:flex'],
     ),
   ).toMatchInlineSnapshot(`
-    "@media (width >= 640px) {
+    ":root, :host {
+      --breakpoint-sm: 640px;
+      --breakpoint-md: 768px;
+      --breakpoint-lg: 1024px;
+      --breakpoint-xl: 1280px;
+      --breakpoint-2xl: 1536px;
+    }
+
+    @media (width >= 640px) {
       .sm\\:flex {
         display: flex;
       }
@@ -807,7 +815,11 @@ test('custom breakpoint', async () => {
       ['10xl:flex'],
     ),
   ).toMatchInlineSnapshot(`
-    "@media (width >= 5000px) {
+    ":root, :host {
+      --breakpoint-10xl: 5000px;
+    }
+
+    @media (width >= 5000px) {
       .\\31 0xl\\:flex {
         display: flex;
       }
@@ -830,7 +842,13 @@ test('max-*', async () => {
       ['max-lg:flex', 'max-sm:flex', 'max-md:flex'],
     ),
   ).toMatchInlineSnapshot(`
-    "@media (width < 1024px) {
+    ":root, :host {
+      --breakpoint-sm: 640px;
+      --breakpoint-lg: 1024px;
+      --breakpoint-md: 768px;
+    }
+
+    @media (width < 1024px) {
       .max-lg\\:flex {
         display: flex;
       }
@@ -879,7 +897,13 @@ test('min-*', async () => {
       ['min-lg:flex', 'min-sm:flex', 'min-md:flex'],
     ),
   ).toMatchInlineSnapshot(`
-    "@media (width >= 640px) {
+    ":root, :host {
+      --breakpoint-sm: 640px;
+      --breakpoint-lg: 1024px;
+      --breakpoint-md: 768px;
+    }
+
+    @media (width >= 640px) {
       .min-sm\\:flex {
         display: flex;
       }
@@ -930,7 +954,15 @@ test('sorting stacked min-* and max-* variants', async () => {
       ['min-sm:max-lg:flex', 'min-sm:max-xl:flex', 'min-md:max-lg:flex', 'min-xs:max-sm:flex'],
     ),
   ).toMatchInlineSnapshot(`
-    "@media (width >= 280px) {
+    ":root, :host {
+      --breakpoint-sm: 640px;
+      --breakpoint-lg: 1024px;
+      --breakpoint-md: 768px;
+      --breakpoint-xl: 1280px;
+      --breakpoint-xs: 280px;
+    }
+
+    @media (width >= 280px) {
       @media (width < 640px) {
         .min-xs\\:max-sm\\:flex {
           display: flex;
@@ -977,7 +1009,13 @@ test('stacked min-* and max-* variants should come after unprefixed variants', a
       ['sm:flex', 'min-sm:max-lg:flex', 'md:flex', 'min-md:max-lg:flex'],
     ),
   ).toMatchInlineSnapshot(`
-    "@media (width >= 640px) {
+    ":root, :host {
+      --breakpoint-sm: 640px;
+      --breakpoint-lg: 1024px;
+      --breakpoint-md: 768px;
+    }
+
+    @media (width >= 640px) {
       .sm\\:flex {
         display: flex;
       }
@@ -1033,7 +1071,13 @@ test('min, max and unprefixed breakpoints', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    "@media (width < 1024px) {
+    ":root, :host {
+      --breakpoint-sm: 640px;
+      --breakpoint-lg: 1024px;
+      --breakpoint-md: 768px;
+    }
+
+    @media (width < 1024px) {
       .max-lg\\:flex {
         display: flex;
       }
@@ -1412,7 +1456,11 @@ test('not', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    ".not-first\\:flex:not(:first-child), .not-last\\:flex:not(:last-child), .not-only\\:flex:not(:only-child), .not-odd\\:flex:not(:nth-child(odd)), .not-even\\:flex:not(:nth-child(2n)), .not-first-of-type\\:flex:not(:first-of-type), .not-last-of-type\\:flex:not(:last-of-type), .not-only-of-type\\:flex:not(:only-of-type), .not-visited\\:flex:not(:visited), .not-target\\:flex:not(:target) {
+    ":root, :host {
+      --breakpoint-sm: 640px;
+    }
+
+    .not-first\\:flex:not(:first-child), .not-last\\:flex:not(:last-child), .not-only\\:flex:not(:only-child), .not-odd\\:flex:not(:nth-child(odd)), .not-even\\:flex:not(:nth-child(2n)), .not-first-of-type\\:flex:not(:first-of-type), .not-last-of-type\\:flex:not(:last-of-type), .not-only-of-type\\:flex:not(:only-of-type), .not-visited\\:flex:not(:visited), .not-target\\:flex:not(:target) {
       display: flex;
     }
 
@@ -1954,7 +2002,11 @@ test('container queries', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    "@container name (width < 1024px) {
+    ":root, :host {
+      --container-lg: 1024px;
+    }
+
+    @container name (width < 1024px) {
       .\\@max-lg\\/name\\:flex {
         display: flex;
       }
@@ -2106,7 +2158,15 @@ test('variant order', async () => {
       ],
     ),
   ).toMatchInlineSnapshot(`
-    "@media (hover: hover) {
+    ":root, :host {
+      --breakpoint-sm: 640px;
+      --breakpoint-md: 768px;
+      --breakpoint-lg: 1024px;
+      --breakpoint-xl: 1280px;
+      --breakpoint-2xl: 1536px;
+    }
+
+    @media (hover: hover) {
       .group-hover\\:flex:is(:where(.group):hover *), .peer-hover\\:flex:is(:where(.peer):hover ~ *) {
         display: flex;
       }